Create README.md
This Site was created as a Midterm Exam for a Web Design course I took. The goal of exam was to take pages we had created in class and make them our own. We were to also add pages that were our own idea. I am currently working on redoing the entire site to improve the design, and making the pages dynamic.

For the exam this was actually a Super Hero fan club, but now that I have gotten farther into the redesign I am considering make it a Comic Book fan club. Before I do so I examining what effect if any this would have on existing functionality.

I am exploring functionality to add such as using the Youtube API in some manner, or allowing members to create their own slideshows. My plan is to also have a administrative tool/dashboard for managing content, which will eventually include submissions by members.